[cryptography] Move DKGs (#3856)
