+++ cryptography/fuzz/Cargo.toml
@@ -133,15 +133,15 @@ doc = false
 bench = false
 
 [[bin]]
-name = "bls12381_dkg"
-path = "fuzz_targets/bls12381_dkg.rs"
+name = "bls12381_dkg_feldman_desmedt"
+path = "fuzz_targets/bls12381_dkg_feldman_desmedt.rs"
 test = false
 doc = false
 bench = false
 
 [[bin]]
-name = "bls12381_golden_dkg"
-path = "fuzz_targets/bls12381_golden_dkg.rs"
+name = "bls12381_dkg_golden"
+path = "fuzz_targets/bls12381_dkg_golden.rs"
 test = false
 doc = false
 bench = false
