+++ docs/blogs/commonware-cryptography.html
@@ -97,7 +97,7 @@
         </ul>
         <p>Today, we're excited to share how @commonwarexyz is answering these questions with <a href="https://docs.rs/commonware-cryptography">commonware-cryptography</a> (now in ALPHA).</p>
         <p>commonware-cryptography is an open (Apache-2 and MIT) implementation of BLS12-381 Distributed Key Generation (DKG), Resharing, and Threshold Signatures in Rust. Unlike most implementations, our construction does not employ a "board" for share distribution (that requires dealers to broadcast encrypted shares over a public channel). Rather, dealers distribute shares over encrypted connections directly with each player and players broadcast an acknowledgement when they receive a correct share (no share material posted).</p>
-        <p>With an eye towards consensus integration (natively emitting Threshold Signatures during finalization), this interactive DKG/Resharing construction requires 2f + 1 players (of 3f + 1 players where any f are Byzantine) to generate a valid threshold signature. Resharing can be run to introduce/remove contributors and/or proactively on a regular interval to bound the age of any shares. You can read more about this construction <a href="https://docs.rs/commonware-cryptography/latest/commonware_cryptography/bls12381/dkg/index.html">here</a>.</p>
+        <p>With an eye towards consensus integration (natively emitting Threshold Signatures during finalization), this interactive DKG/Resharing construction requires 2f + 1 players (of 3f + 1 players where any f are Byzantine) to generate a valid threshold signature. Resharing can be run to introduce/remove contributors and/or proactively on a regular interval to bound the age of any shares. You can read more about this construction <a href="https://docs.rs/commonware-cryptography/latest/commonware_cryptography/bls12381/dkg/feldman_desmedt/index.html">here</a>.</p>
         <p>So, how does this answer any of a developer's questions?</p>
         <ul>
             <li>
