+++ docs/blogs/golden.html
@@ -535,7 +535,7 @@ Meier</a></div>
         <h1 id="performance">Performance</h1>
         <p>Here are the current performance numbers we have from our
         initial implementation. Measured values are taken from
-        <code>bls12381::golden_dkg::{deal, verify, dealing_size}</code>
+        <code>bls12381::dkg::golden::{deal, verify, dealing_size}</code>
         for <span class="math inline">n \in \{2, 4, 8\}</span>; values
         for <span class="math inline">n \in \{16, 32, 64\}</span> are
         projected linearly (shown in gray) since the underlying cost is
